WebSocketSession unit tests
