+++ Sources/WalletConnect_Swift/Networking/URLSession+Protocols.swift
@@ -1,0 +1,20 @@
+import Foundation
+
+protocol URLSessionProtocol {
+    func webSocketTask(with url: URL) -> URLSessionWebSocketTaskProtocol
+}
+
+extension URLSession: URLSessionProtocol {
+    func webSocketTask(with url: URL) -> URLSessionWebSocketTaskProtocol {
+        webSocketTask(with: url) as URLSessionWebSocketTask
+    }
+}
+
+protocol URLSessionWebSocketTaskProtocol {
+    func resume()
+    func cancel()
+    func send(_ message: URLSessionWebSocketTask.Message, completionHandler: @escaping (Error?) -> Void)
+    func receive(completionHandler: @escaping (Result<URLSessionWebSocketTask.Message, Error>) -> Void)
+}
+
+extension URLSessionWebSocketTask: URLSessionWebSocketTaskProtocol {}
