+++ Tests/WalletConnect_SwiftTests/Helpers/URL+Extension.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension URL {
+    
+    static func stub() -> URL {
+        URL(string: "https://httpbin.org")!
+    }
+}
